Restructure Cucumber test packages again
