+++ src/test/kotlin/de/livepoll/api/cucumber/CucumberIntegrationTestsGlueSetup.kt
@@ -1,4 +1,4 @@
-package de.livepoll.api.cucumber.poll
+package de.livepoll.api.cucumber
 
 import io.cucumber.junit.Cucumber
 import io.cucumber.junit.CucumberOptions
@@ -7,8 +7,8 @@ import org.junit.runner.RunWith
 
 @RunWith(Cucumber::class)
 @CucumberOptions(
-        features = ["src/test/resources/features/poll.feature"],
-        plugin = ["pretty", "html:target/cucumber"])
-class PollCucumberIntegrationTest {
+        features = ["src/test/resources/features/"],
+        plugin = ["pretty", "html:target/cucumber-report.html"])
+class CucumberIntegrationTestsGlueSetup {
 
 }
